# api/openapi.yaml
openapi: 3.1.0
info:
  title: buffgo local control API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/targets:
    get:
      operationId: listTargets
      responses:
        "200":
          description: Summary collection targets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Target"
    post:
      operationId: createTarget
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetCreate"
      responses:
        "201":
          description: Target created or existing identical target returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /api/targets/{id}/desired:
    parameters:
      - $ref: "#/components/parameters/TargetID"
    post:
      operationId: setTargetDesired
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetDesired"
      responses:
        "200":
          description: Desired state applied; actual state follows scheduler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/targets/{id}/sort:
    parameters:
      - $ref: "#/components/parameters/TargetID"
    post:
      operationId: setTargetSortOrder
      description: >-
        Changes the platform ordering this target collects with. The cursor is a
        list offset, so this voids the active batch and the next cycle restarts
        from the beginning. Stored market facts are kept.
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TargetSortOrder"
      responses:
        "200":
          description: Ordering applied; the active batch is voided
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Target"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/targets/{id}/delete:
    parameters:
      - $ref: "#/components/parameters/TargetID"
    post:
      operationId: deleteTarget
      description: Only a disabled and stopped target that never ran can be deleted; collection history is never removed
      security:
        - controlSession: []
          csrfToken: []
      responses:
        "200":
          description: Target deleted
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/runs:
    get:
      operationId: listRuns
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        "200":
          description: Recent collection runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
...
  /api/quotes:
    get:
      operationId: listQuotes
      parameters:
        - name: appid
          in: query
          schema: {type: integer, format: int64, minimum: 1}
        - name: platform
          in: query
          schema: {type: string}
        - name: side
          in: query
          schema: {type: string, enum: [bid, ask]}
        - name: keyword
          in: query
          description: Case-insensitive substring match on the product name.
          schema: {type: string, maxLength: 128}
        - name: item_type
          in: query
          schema: {type: string, maxLength: 128}
        - name: min_cents
          in: query
          description: Only products with a last present price at or above this value.
          schema: {type: integer, format: int64, minimum: 0}
        - name: max_cents
          in: query
          schema: {type: integer, format: int64, minimum: 0}
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_desc, price_asc, listings_desc, name]
            default: price_desc
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 60}
      responses:
        "200":
          description: One page of latest attempts with the total match count
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [total, quotes]
                properties:
                  total: {type: integer, format: int64, minimum: 0}
                  quotes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Quote"
components:
  parameters:
...
    TargetID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  securitySchemes:
    controlSession:
      type: apiKey
      in: cookie
      name: buffgo_control_session
    csrfToken:
      type: apiKey
      in: header
      name: X-Buffgo-CSRF
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - invalid_host
            - method_not_allowed
            - origin_rejected
            - origin_required
            - json_required
            - csrf_rejected
            - internal_error
            - invalid_account_id
            - account_not_found
            - account_occupied
            - account_in_use
            - account_revision_conflict
            - session_required
            - session_too_large
            - invalid_json
            - body_not_allowed
            - resource_storage_unavailable
            - resource_integrity
            - resource_error
            - not_found
            - invalid_account
            - account_conflict
            - invalid_node_id
            - node_not_found
            - node_occupied
            - node_in_use
            - node_revision_conflict
            - node_assignment_conflict
            - node_name_conflict
            - platform_region_mismatch
            - invalid_node
            - invalid_exit_address
            - invalid_combination_id
            - combination_not_found
            - combination_occupied
            - combination_incompatible
            - combination_conflict
            - invalid_combination
            - resource_in_use
            - credential_required
            - credential_too_large
            - invalid_provider_id
            - provider_not_found
            - provider_revision_conflict
            - provider_conflict
            - invalid_provider
            - invalid_target_id
            - target_not_found
            - collection_not_found
            - collection_conflict
            - target_not_stopped
            - collection_in_use
            - invalid_collection
            - collection_storage_unavailable
            - collection_error
            - invalid_run_id
            - invalid_page_sequence
            - page_payload_not_found
            - invalid_limit
            - invalid_offset
            - invalid_price_range
            - invalid_appid
            - invalid_quotes
...
    Target:
      type: object
      additionalProperties: false
      required: [id, revision, platform, appid, side, desired, actual, switch_version, changed_at]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        revision: {type: integer, format: int64, minimum: 1}
        platform: {type: string}
        appid: {type: integer, format: int64, minimum: 1}
        side: {type: string, enum: [bid, ask]}
        desired: {type: string, enum: [enabled, disabled]}
        actual: {type: string, enum: [starting, waiting, running, blocked, stopping, stopped, error]}
        switch_version: {type: integer, format: int64, minimum: 1}
        reason: {type: string}
        recovery: {type: string}
        sort_column:
          type: string
          enum: [price, quantity, name]
          description: Platform ordering this target collects with.
        sort_dir: {type: string, enum: [asc, desc]}
        recheck_at: {type: string, format: date-time}
        changed_at: {type: string, format: date-time}
    TargetCreate:
      type: object
      additionalProperties: false
      required: [platform, appid, side]
      properties:
        platform: {type: string}
        appid: {type: integer, format: int64, minimum: 1}
        side: {type: string, enum: [bid, ask]}
        desired: {type: string, enum: [enabled, disabled]}
    TargetDesired:
      type: object
      additionalProperties: false
      required: [expected_revision, desired]
      properties:
        expected_revision: {type: integer, format: int64, minimum: 1}
        desired: {type: string, enum: [enabled, disabled]}
    TargetSortOrder:
      type: object
      additionalProperties: false
      required: [expected_revision, sort_column, sort_dir]
      properties:
        expected_revision: {type: integer, format: int64, minimum: 1}
        sort_column: {type: string, enum: [price, quantity, name]}
        sort_dir: {type: string, enum: [asc, desc]}
    Run:
      type: object
      additionalProperties: false
      required: [id, task_type, platform, appid, state, last_page_sequence, created_at]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        target_id: {type: integer, format: int64, minimum: 1}
        task_type: {type: string, enum: [summary, detail]}
        platform: {type: string}
        appid: {type: integer, format: int64, minimum: 1}
        side: {type: string, enum: [bid, ask]}
        state: {type: string, enum: [pending, running, succeeded, failed, stopped]}
        completeness: {type: string, enum: [complete, partial]}
        reason: {type: string}
        cursor: {type: string}
        last_page_sequence: {type: integer, format: int64, minimum: 0}
        created_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
...
    Quote:
      type: object
      additionalProperties: false
      required: [product_id, appid, name, platform, side, status, collected_at]
      properties:
        product_id: {type: integer, format: int64, minimum: 1}
        appid: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        icon_path:
          type: string
          maxLength: 512
          description: >-
            Platform image path fragment, never a full URL and never containing
            a slash. Clients prepend the Steam economy image CDN prefix.
        item_type:
          type: string
          maxLength: 128
          description: Platform category. Absent for games that do not report one.
        name_color:
          type: string
          pattern: "^[0-9a-fA-F]{6}$"
          description: Quality color without a leading hash.
        platform: {type: string}
        side: {type: string, enum: [bid, ask]}
        status: {type: string, enum: [present, empty, unavailable, failed]}
        reason_code: {type: string}
        collected_at: {type: string, format: date-time}
        source_time: {type: string, format: date-time}
        present_cents: {type: integer, format: int64, minimum: 0}
        present_order_count: {type: integer, format: int64, minimum: 0}
        present_collected_at: {type: string, format: date-time}
  responses:
    Error:
      description: Request rejected without reflecting request data or credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"